// type.d.ts
export interface productProps {
  category: string;
  description: string;
  id: number;
  image: string;
  price: number;
  rating: {
    count: number;
    rate: number;
  };
  title: string;
}
export interface sliceData {
  category: string;
  description: string;
  id: number;
  image: string;
  price: number;
  rating: {
    count: number;
    rate: number;
  };
  title: string;
  quantity: number;
}
export interface stateProps{
  productData;[],
  favouriteData: [],
  userLogin: null|string,
  next:any
};
export interface dropDown {
  category: string;
  description: string;
  id: number;
  image: string;
  price: number;
  rating: {
    count: number;
    rate: number;
  };
  title: string;
  quantity: number;
}